+++ common/src/main/java/xyz/rrtt217/HDRMod/mixin/MixinDebugScreenEntryList.java
@@ -1,0 +1,25 @@
+package xyz.rrtt217.HDRMod.mixin;
+
+import net.minecraft.client.gui.components.debug.DebugScreenEntryList;
+import net.minecraft.client.gui.components.debug.DebugScreenEntryStatus;
+import net.minecraft.resources.ResourceLocation;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import java.util.Map;
+
+@Mixin(DebugScreenEntryList.class)
+public class MixinDebugScreenEntryList {
+    @Shadow
+    private Map<ResourceLocation, DebugScreenEntryStatus> allStatuses;
+
+    @Inject(method = "rebuildCurrentList", at = @At("HEAD"))
+    private void hdr_mod$addHDRModDebugEntries(CallbackInfo ci) {
+        if (!this.allStatuses.containsKey(ResourceLocation.fromNamespaceAndPath("hdr_mod", "debug"))) {
+            this.allStatuses.put(ResourceLocation.fromNamespaceAndPath("hdr_mod", "debug"), DebugScreenEntryStatus.NEVER);
+        }
+    }
+}