+++ common/src/main/java/xyz/rrtt217/HDRMod/mixin/MixinRenderSystem.java
@@ -10,14 +10,13 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import xyz.rrtt217.HDRMod.util.SetupBeforeGLFWInit;
 
-import static xyz.rrtt217.HDRMod.HDRMod.enableHDR;
 
 @Mixin(value = RenderSystem.class, priority = 1010)
 public class MixinRenderSystem {
     // Enables Wayland color management. No effect on non-wayland platform.
     @Inject(method = "initBackendSystem", at = @At("HEAD"))
     private static void hdr_mod$colorManagementHint(CallbackInfoReturnable<TimeSource.NanoTimeSource> cir) {
-        if(enableHDR && HDRModMixinPlugin.hasGlfwLib) {
+        if(HDRModMixinPlugin.hasGlfwLib) {
             SetupBeforeGLFWInit.setup();
             GLFW.glfwInitHint(0x00026002,GLFW.GLFW_TRUE);
             if(Platform.isWindows()) {